+++ FrontEnd/lib/forum/screens/Post.dart
@@ -1,8 +1,9 @@
 import 'dart:html';
 
 import 'package:flutter/material.dart';
-import 'package:homepage/forum/screens/Utilities.dart';
+import 'package:homepage/forum/utilities/Utilities.dart';
 import 'package:homepage/forum/classes/PostClass.dart';
+import 'package:intl/intl.dart';
 
 class POST extends StatefulWidget {
   const POST({
@@ -71,9 +72,9 @@ class _POSTState extends State<POST> {
                   Padding(
                     //AUTHOR + DATE
                     padding: EdgeInsets.only(
-                      left: 35,
-                      right: 35,
-                      bottom: 35,
+                      left: 20,
+                      right: 20,
+                      bottom: 20,
                     ),
                     child: Container(
                         decoration: BoxDecoration(
@@ -85,8 +86,8 @@ class _POSTState extends State<POST> {
                           children: [
                             Padding(
                               padding: EdgeInsets.only(
-                                left: 35,
-                                right: 35,
+                                left: 20,
+                                right: 20,
                                 bottom: 15,
                                 top: 20,
                               ),
@@ -97,7 +98,13 @@ class _POSTState extends State<POST> {
                                       child: Row(
                                         children: [
                                           Icon(Icons.person),
-                                          Text(widget.post.author),
+                                          Text(
+                                            widget.post.author,
+                                            style: TextStyle(
+                                              fontWeight: FontWeight.bold,
+                                              fontSize: 14.0,
+                                            ),
+                                          ),
                                         ],
                                       )),
                                   Expanded(
@@ -105,19 +112,28 @@ class _POSTState extends State<POST> {
                                       child: Row(
                                         children: [
                                           Icon(Icons.calendar_today),
-                                          Text(widget.post.date),
+                                          Text(
+                                            widget.post.date,
+                                            style: TextStyle(
+                                              fontWeight: FontWeight.bold,
+                                              fontSize: 14.0,
+                                            ),
+                                          ),
                                         ],
                                       )),
                                 ],
                               ),
                             ),
-                            ContentBox(35,
-                                35), //POST TEXT + COMMENTS AND LIKEBUTTON SHAREBUTTON REPORTBUTTON
+                            ContentBox(
+                              20,
+                              20,
+                              13,
+                            ), //POST TEXT + COMMENTS AND LIKEBUTTON SHAREBUTTON REPORTBUTTON
                             Padding(
                               padding: EdgeInsets.only(
-                                left: 35,
-                                right: 35,
-                                bottom: 35,
+                                left: 20,
+                                right: 20,
+                                bottom: 20,
                               ),
                               child: ConstrainedBox(
                                 constraints: new BoxConstraints(
@@ -134,15 +150,26 @@ class _POSTState extends State<POST> {
                                       children: [
                                         Container(
                                           padding: EdgeInsets.only(
-                                              top: 35 / 2,
-                                              left: 35,
-                                              right: 35,
-                                              bottom: 35 / 2),
+                                              top: 20 / 2,
+                                              left: 20,
+                                              right: 20,
+                                              bottom: 20 / 2),
                                           child: Column(
                                             children: [
                                               Expanded(
-                                                  flex: 0,
-                                                  child: Text("Comments:")),
+                                                flex: 0,
+                                                child: Text(
+                                                  "Comments:",
+                                                  style: TextStyle(
+                                                    fontWeight: FontWeight.bold,
+                                                    fontSize: 15.0,
+                                                  ),
+                                                ),
+                                              ),
+                                              Expanded(
+                                                flex: 0,
+                                                child: SizedBox(height: 5),
+                                              ),
                                               Expanded(
                                                 flex: 0,
                                                 child:
@@ -151,26 +178,115 @@ class _POSTState extends State<POST> {
                                             ],
                                           ),
                                         ),
-                                        CommentBox(
-                                          35,
-                                          35,
-                                          widget.post.comments[0].author,
-                                          widget.post.comments[0].content,
-                                          widget.post.comments[0].date,
-                                        ),
-                                        CommentBox(
-                                          35,
-                                          35,
-                                          widget.post.comments[1].author,
-                                          widget.post.comments[1].content,
-                                          widget.post.comments[1].date,
-                                        ),
-                                        CommentBox(
-                                          35,
-                                          35,
-                                          widget.post.comments[2].author,
-                                          widget.post.comments[2].content,
-                                          widget.post.comments[2].date,
+                                        ListView.builder(
+                                          scrollDirection: Axis.vertical,
+                                          shrinkWrap: true,
+                                          itemCount:
+                                              widget.post.comments.length,
+                                          itemBuilder: (context, index) {
+                                            return Container(
+                                              padding: EdgeInsets.only(
+                                                left: 20,
+                                                right: 20,
+                                                bottom: 20 / 2,
+                                              ),
+                                              child: ConstrainedBox(
+                                                constraints: new BoxConstraints(
+                                                  minWidth: 700,
+                                                ),
+                                                child: Container(
+                                                  decoration: BoxDecoration(
+                                                    borderRadius:
+                                                        BorderRadius.all(
+                                                      Radius.circular(20),
+                                                    ),
+                                                    color: Color(0xFFD9F7FF),
+                                                  ),
+                                                  child: Column(children: [
+                                                    Padding(
+                                                      padding: EdgeInsets.only(
+                                                        left: 20,
+                                                        right: 20,
+                                                        bottom: 20 / 2,
+                                                        top: 0.3 * 20,
+                                                      ),
+                                                      child: Column(
+                                                        mainAxisAlignment:
+                                                            MainAxisAlignment
+                                                                .spaceEvenly,
+                                                        children: [
+                                                          Expanded(
+                                                              flex: 0,
+                                                              child: Row(
+                                                                children: [
+                                                                  Icon(Icons
+                                                                      .person),
+                                                                  Text(
+                                                                    widget
+                                                                        .post
+                                                                        .comments[
+                                                                            index]
+                                                                        .author,
+                                                                    style:
+                                                                        TextStyle(
+                                                                      fontWeight:
+                                                                          FontWeight
+                                                                              .bold,
+                                                                      fontSize:
+                                                                          14.0,
+                                                                    ),
+                                                                  ),
+                                                                ],
+                                                              )),
+                                                          Expanded(
+                                                              flex: 0,
+                                                              child: Row(
+                                                                children: [
+                                                                  Icon(Icons
+                                                                      .calendar_today),
+                                                                  Text(
+                                                                    widget
+                                                                        .post
+                                                                        .comments[
+                                                                            index]
+                                                                        .date,
+                                                                    style:
+                                                                        TextStyle(
+                                                                      fontWeight:
+                                                                          FontWeight
+                                                                              .bold,
+                                                                      fontSize:
+                                                                          14.0,
+                                                                    ),
+                                                                  ),
+                                                                ],
+                                                              )),
+                                                        ],
+                                                      ),
+                                                    ),
+                                                    Container(
+                                                      padding: EdgeInsets.only(
+                                                        left: 20,
+                                                        right: 20,
+                                                        bottom: 20,
+                                                      ),
+                                                      child: Text(
+                                                        widget
+                                                            .post
+                                                            .comments[index]
+                                                            .content,
+                                                        style: TextStyle(
+                                                          fontWeight:
+                                                              FontWeight.bold,
+                                                          fontSize: 13.0,
+                                                        ),
+                                                      ),
+                                                    )
+                                                  ]),
+                                                ),
+                                              ),
+                                            );
+                                          },
                                         ),
                                       ],
                                     )),
@@ -213,7 +329,7 @@ class _POSTState extends State<POST> {
                       ],
                     ),
                   ),
-                  PostBox(100, 40, 40),
+                  PostBox(100, 40, 40, 14, 13),
                 ],
               ),
             );
@@ -245,7 +361,7 @@ class _POSTState extends State<POST> {
                       ],
                     ),
                   ),
-                  PostBox(150, 50, 50),
+                  PostBox(150, 50, 50, 15, 14),
                 ],
               ),
             );
@@ -257,7 +373,10 @@ class _POSTState extends State<POST> {
 
 //------------------------------------------------------- METHODS --------------------------------------------------------------------------------------------
 
-  Padding PostBox(double outPadding, double inPadding, double TextPadding) {
+//-------------------------------------------------------------------------------------- POST BOX --------------------------------------------------------------------------------------------
+
+  Padding PostBox(double outPadding, double inPadding, double TextPadding,
+      double MainFontSize, double SecondayFontSize) {
     return Padding(
       //AUTHOR + DATE
       padding: EdgeInsets.only(
@@ -273,17 +392,19 @@ class _POSTState extends State<POST> {
         alignment: Alignment.center,
         child: Column(
           children: [
-            AuthorAndDate(inPadding, widget.post.author, widget.post.date),
-            ContentBox(inPadding, TextPadding),
-            CommentsBox(inPadding,
-                TextPadding), //POST TEXT + COMMENTS AND LIKEBUTTON SHAREBUTTON REPORTBUTTON
+            AuthorAndDate(
+                inPadding, widget.post.author, widget.post.date, MainFontSize),
+            ContentBox(inPadding, TextPadding, SecondayFontSize),
+            CommentsBox(inPadding, TextPadding,
+                SecondayFontSize), //POST TEXT + COMMENTS AND LIKEBUTTON SHAREBUTTON REPORTBUTTON
           ],
         ),
       ),
     );
   }
 
-  Padding AuthorAndDate(double inPadding, String author, String date) {
+  Padding AuthorAndDate(
+      double inPadding, String author, String date, double MainFontSize) {
     return Padding(
       padding: EdgeInsets.only(
         left: inPadding,
@@ -299,7 +420,13 @@ class _POSTState extends State<POST> {
               child: Row(
                 children: [
                   Icon(Icons.person),
-                  Text(author),
+                  Text(
+                    author,
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: MainFontSize,
+                    ),
+                  ),
                 ],
               )),
           Expanded(
@@ -307,7 +434,13 @@ class _POSTState extends State<POST> {
               child: Row(
                 children: [
                   Icon(Icons.calendar_today),
-                  Text(date),
+                  Text(
+                    date,
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: MainFontSize,
+                    ),
+                  ),
                 ],
               )),
         ],
@@ -315,7 +448,8 @@ class _POSTState extends State<POST> {
     );
   }
 
-  Padding ContentBox(double interiorPadding, double textPadding) {
+  Padding ContentBox(
+      double interiorPadding, double textPadding, double FontSize) {
     return Padding(
       padding: EdgeInsets.only(
         left: interiorPadding,
@@ -342,6 +476,10 @@ class _POSTState extends State<POST> {
                 Container(
                   child: Text(
                     widget.post.content,
+                    style: TextStyle(
+                      fontWeight: FontWeight.bold,
+                      fontSize: FontSize + 1,
+                    ),
                   ),
                   alignment: Alignment.topLeft,
                 ),
@@ -376,7 +514,9 @@ class _POSTState extends State<POST> {
     );
   }
 
-  Padding CommentsBox(double inPadding, double textPadding) {
+  //-------------------------------------------------------------------------------------- COMMENTS BOX --------------------------------------------------------------------------------------------
+
+  Padding CommentsBox(double inPadding, double textPadding, double FontSize) {
     return Padding(
       padding: EdgeInsets.only(
         left: inPadding,
@@ -388,59 +528,74 @@ class _POSTState extends State<POST> {
           minWidth: 1000,
         ),
         child: Container(
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.all(Radius.circular(20)),
-              color: Color(0xFFAAECF5),
-            ),
-            alignment: Alignment.center,
-            child: Column(
-              children: [
-                Container(
-                  padding: EdgeInsets.only(
-                      top: textPadding / 2,
-                      left: textPadding,
-                      right: textPadding,
-                      bottom: textPadding / 2),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                    children: [
-                      Expanded(flex: 1, child: Text("Comments:")),
-                      Expanded(
-                        flex: 0,
-                        child: ButtonPostComment(150, 30),
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.all(Radius.circular(20)),
+            color: Color(0xFFAAECF5),
+          ),
+          alignment: Alignment.center,
+          child: Column(
+            children: [
+              Container(
+                padding: EdgeInsets.only(
+                    top: textPadding / 2,
+                    left: textPadding,
+                    right: textPadding,
+                    bottom: textPadding / 2),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                  children: [
+                    Expanded(
+                        flex: 1,
+                        child: Text(
+                          "Comments:",
+                          style: TextStyle(
+                            fontWeight: FontWeight.bold,
+                            fontSize: 18.0,
+                          ),
+                        )),
+                    Expanded(
+                      flex: 0,
+                      child: ButtonPostComment(150, 30),
+                    ),
+                  ],
+                ),
+              ),
+              ListView.builder(
+                scrollDirection: Axis.vertical,
+                shrinkWrap: true,
+                itemCount: widget.post.comments.length,
+                itemBuilder: (context, index) {
+                  if (widget.post.comments.isNotEmpty) {
+                    return CommentBox(
+                      inPadding,
+                      textPadding,
+                      widget.post.comments[index].author,
+                      widget.post.comments[index].content,
+                      widget.post.comments[index].date,
+                      FontSize,
+                    );
+                  } else {
+                    return Text(
+                      "No comments...",
+                      style: TextStyle(
+                        fontWeight: FontWeight.bold,
+                        fontSize: FontSize,
                       ),
-                    ],
-                  ),
-                ),
-                CommentBox(
-                  inPadding,
-                  textPadding,
-                  widget.post.comments[0].author,
-                  widget.post.comments[0].content,
-                  widget.post.comments[0].date,
-                ),
-                CommentBox(
-                  inPadding,
-                  textPadding,
-                  widget.post.comments[1].author,
-                  widget.post.comments[1].content,
-                  widget.post.comments[1].date,
-                ),
-                CommentBox(
-                  inPadding,
-                  textPadding,
-                  widget.post.comments[2].author,
-                  widget.post.comments[2].content,
-                  widget.post.comments[2].date,
-                ),
-              ],
-            )),
+                    );
+                  }
+                },
+              ),
+            ],
+          ),
+        ),
       ),
     );
   }
 
+  //-------------------------------------------------------------------------------------- COMMENT BOX --------------------------------------------------------------------------------------------
+
   Container CommentBox(double inPadding, double textPadding, String author,
-      String content, String date) {
+      String content, String date, double FontSize) {
     return Container(
       padding: EdgeInsets.only(
         left: inPadding,
@@ -463,6 +618,7 @@ class _POSTState extends State<POST> {
               textPadding,
               author,
               date,
+              FontSize + 1,
             ),
             Container(
               padding: EdgeInsets.only(
@@ -470,13 +626,21 @@ class _POSTState extends State<POST> {
                 right: textPadding,
                 bottom: textPadding,
               ),
-              child: Text(content),
+              child: Text(
+                content,
+                style: TextStyle(
+                  fontWeight: FontWeight.bold,
+                  fontSize: FontSize,
+                ),
+              ),
             )
           ]),
         ),
       ),
     );
   }
+
+  //-------------------------------------------------------------------------------------- BUTTON LIKE --------------------------------------------------------------------------------------------
 
   Container ButtonLike(double height, double width) {
     return Container(
@@ -509,6 +673,8 @@ class _POSTState extends State<POST> {
     );
   }
 
+  //-------------------------------------------------------------------------------------- BUTTON SHARE --------------------------------------------------------------------------------------------
+
   Container ButtonShare(double height, double width) {
     return Container(
       height: height,
@@ -528,6 +694,8 @@ class _POSTState extends State<POST> {
     );
   }
 
+  //-------------------------------------------------------------------------------------- BUTTON POST COMMENT --------------------------------------------------------------------------------------------
+
   Container ButtonPostComment(double width, double height) {
     return Container(
       height: height,
@@ -543,8 +711,11 @@ class _POSTState extends State<POST> {
           if (comment == null || comment.isEmpty) {
             return;
           }
-          setState(() => widget.post
-              .addComment(1, widget.post.id, comment, "author", "date", 0));
+          var now = new DateTime.now();
+          var formatter = DateFormat('dd/MM/yyyy');
+          String formattedDate = formatter.format(now);
+          setState(() => widget.post.addComment(
+              1, 1, widget.post.id, comment, "author", formattedDate, 0));
           controller.clear();
         },
       ),
@@ -552,29 +723,30 @@ class _POSTState extends State<POST> {
   }
 
   Future<String?> OpenDialog() => showDialog<String>(
-      context: context,
-      builder: (context) => AlertDialog(
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(30),
+        context: context,
+        builder: (context) => AlertDialog(
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(30),
+          ),
+          title: Text('Type your comment'),
+          content: TextField(
+            autofocus: true,
+            keyboardType: TextInputType.multiline,
+            minLines: 1,
+            maxLines: 20,
+            decoration: InputDecoration(
+              hintText: 'Enter your comment here',
             ),
-            title: Text('Type your comment'),
-            content: TextField(
-              autofocus: true,
-              keyboardType: TextInputType.multiline,
-              minLines: 1,
-              maxLines: 20,
-              decoration: InputDecoration(
-                hintText: 'Enter your comment here',
-              ),
-              controller: controller,
+            controller: controller,
+          ),
+          actions: [
+            TextButton(
+              onPressed: () {
+                Navigator.of(context).pop(controller.text);
+              },
+              child: Text('Submit'),
             ),
-            actions: [
-              TextButton(
-                onPressed: () {
-                  Navigator.of(context).pop(controller.text);
-                },
-                child: Text('Submit'),
-              ),
-            ],
-          ));
+          ],
+        ),
+      );
 }
