+++ FrontEnd/lib/forum/utilities/Utilities.dart
@@ -1,0 +1,9 @@
+export 'buttons/ButtonBack.dart';
+export 'buttons/ButtonMyPosts.dart';
+export 'buttons/ButtonNewPost.dart';
+
+export 'CustomTitle.dart';
+export 'GradientText.dart';
+export 'PostsTable.dart';
+export 'SearchBar.dart';
+export 'Sort.dart';
