+++ FrontEnd/pubspec.yaml
@@ -29,12 +29,14 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  intl: ^0.16.1
+  universal_io: ^2.0.4
     
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
- # http: ^0.12.0+1
+  http: 0.13.4
   simple_gradient_text: ^1.2.2
   flame: ^1.1.1
